backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
gunicorn==21.2.0
python-dotenv==1.0.1

# Security & Cryptography
cryptography==42.0.2
bcrypt==4.1.2
PyJWT==2.8.1
python-jose==3.3.0
passlib==1.7.4
pyOpenSSL==24.0.0
paramiko==3.4.0
cryptography-x509==0.6.0

# Media Processing
Pillow==10.2.0
piexif==1.1.3
moviepy==1.0.3
opencv-python==4.9.0.80

# File Handling
python-magic==0.4.27
aiofiles==23.2.1
scp==0.14.5
shortuuid==1.0.11
pathlib==1.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
validators==0.22.0
python-slugify==8.0.1
colorama==0.4.6  # For colored terminal output

# CORS
fastapi-cors==0.0.6

# QR Code
qrcode==7.4.2
python-barcode==0.14.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0